Fix incorrect address size

+++ src/Memory/Memory.cpp
@@ -1,6 +1,6 @@
 #include "Memory.h"
 
-std::uint8_t Memory::getByte(std::uint8_t address) {
+std::uint8_t Memory::getByte(std::uint16_t address) {
     return memory.at(address);
 }
